ci(mlu): upgrade toolkit image for mlu (#229)
* ci(mlu): upgrade toolkit image for mlu

* ci(test): fix the data type for serial_matmul test

---------

+++ tests/mlu/test_fuse_serial_matmul.py
@@ -8,7 +8,7 @@ from xpu_graph.test_utils import is_similar
 
 aten = torch.ops.aten
 device = "mlu:0"
-dtype = torch.bfloat16
+dtype = torch.float16
 
 from xpu_graph.config import OptLevel
 from xpu_graph.test_utils import (
